refactor: restructure handlers/ into domain-based subpackages
Move 23 handler files into 7 domain subpackages:
- core/ (auth, common, fallback, registration)
- company/ (company, groups, objects)
- users/ (admin, user_activation, user_edit, role_permissions)
- tests/ (tests, test_taking, broadcast)
- training/ (learning_paths, trajectory_editor, trainee_trajectory, mentorship, mentor_assignment)
- management/ (manager_attestation, manager_menu, employee_transition)
- knowledge/ (knowledge_base)

Update all cross-imports and main.py. Router registration order unchanged.
