+++ handlers/company/__init__.py
@@ -1,0 +1,1 @@
+from handlers.company import company, groups, objects
