+++ handlers/core/__init__.py
@@ -1,0 +1,1 @@
+from handlers.core import auth, common, fallback, registration
