+++ handlers/knowledge/__init__.py
@@ -1,0 +1,1 @@
+from handlers.knowledge import knowledge_base
