+++ handlers/tests/__init__.py
@@ -1,0 +1,1 @@
+from handlers.tests import tests, test_taking, broadcast
