+++ handlers/training/__init__.py
@@ -1,0 +1,1 @@
+from handlers.training import learning_paths, trajectory_editor, trainee_trajectory, mentorship, mentor_assignment
