+++ handlers/users/__init__.py
@@ -1,0 +1,1 @@
+from handlers.users import admin, user_activation, user_edit, role_permissions
