+++ handlers/users/role_permissions.py
@@ -16,7 +16,7 @@ from keyboards.keyboards import (
 )
 from states.states import AdminStates
 from utils.logger import log_user_action, log_user_error
-from handlers.auth import check_auth
+from handlers.core.auth import check_auth
 
 router = Router()
