+++ handlers/users/user_activation.py
@@ -18,7 +18,7 @@ from keyboards.keyboards import get_main_menu_keyboard, get_new_users_list_keybo
 from states.states import UserActivationStates
 from utils.logger import log_user_action, log_user_error
 from utils.bot_commands import set_bot_commands
-from handlers.auth import check_auth
+from handlers.core.auth import check_auth
 
 router = Router()
